+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ### Unreleased
 
+- Ensure docker compose configs are always generated relative to the kubetools config file
+
+
 # v12.0.2
 
 - Fix issue when creating new deployments alongside existing ones